changed baseAmount back to a double value

+++ SimpleCurrency/AppDelegate.swift
@@ -12,7 +12,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         UINavigationBar.appearance().barStyle = .black
         if lastUsedBuild == 0 {
             UserDefaults.standard.set("USD", forKey: "baseCurrency")
-            UserDefaults.standard.set(1 as Int, forKey: "baseAmount")
+            UserDefaults.standard.set(1.0, forKey: "baseAmount")
             let currencies: [String] = [Currencies.EUR, .JPY, .GBP, .AUD, .CAD].map { $0.currencyCode }
             UserDefaults.standard.set(currencies, forKey: "currencies")
         }
